change pressure unit hPa to Pa, update print and wifi functions, update docs

+++ esp8266/WetterLib.h
@@ -17,6 +17,6 @@ struct WeatherData {
 	float rainfallD;
 	/// @brief Humidity (%)
 	int humidity;
-	/// @brief Pressure, barometric (hPa)
-	float pressure;
+	/// @brief Pressure (Pa)
+	int pressure;
 };
